// src/app/shared/models/section2-model.ts
import { Section0Model } from "./section0-model";

export class Section2Model extends Section0Model {
  consInfoPep?: number = 0;
  orgInternPregunta1: string = "";
  cargoPoliPregunta2: string = "";
  orgInternPregunta3: string = "";
  orgInternPregunta4: string = "";
  descCargoPublico: number = 0;
  fechaInicio: 1538424211000;
  fechaFin: 1540238620000;
  actualmente: number = 0;
}